Separate starish additions as much as possible
