+++ starish/starishlib.go
@@ -1,24 +1,25 @@
-package starlark
+package starish
 
 import (
 	"fmt"
 
 	"github.com/kassybas/shell-exec/exec"
+	"go.starlark.net/starlark"
 )
 
-func sh(thread *Thread, b *Builtin, args Tuple, kwargs []Tuple) (Value, error) {
+func Sh(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
 	shellPath := "/bin/sh"
 	sep := "_"
 	silent := false
 	shieldEnv := false
-	var env *Dict
-	if err := UnpackArgs("sh", nil, kwargs, "shell?", &shellPath, "shield_env?", &shieldEnv, "silent?", &silent, "sep?", &sep, ":env?", &env); err != nil {
+	var env *starlark.Dict
+	if err := starlark.UnpackArgs("sh", nil, kwargs, "shell?", &shellPath, "shield_env?", &shieldEnv, "silent?", &silent, "sep?", &sep, ":env?", &env); err != nil {
 		return nil, err
 	}
 	if len(args) != 1 {
 		return nil, fmt.Errorf("sh: got %d arguments, want exactly 1", len(args))
 	}
-	script, ok := AsString(args.Index(0))
+	script, ok := starlark.AsString(args.Index(0))
 	if !ok {
 		return nil, fmt.Errorf("sh: non-string argument %+v, want string", args.Index(0))
 	}
@@ -36,18 +37,18 @@ func sh(thread *Thread, b *Builtin, args Tuple, kwargs []Tuple) (Value, error) {
 	if err != nil {
 		return nil, err
 	}
-	result := []Value{String(output), String(errOutput), MakeInt(statusCode)}
-	return NewList(result), nil
+	result := []starlark.Value{starlark.String(output), starlark.String(errOutput), starlark.MakeInt(statusCode)}
+	return starlark.NewList(result), nil
 }
 
-func getEnvVars(arg Value, sep string) ([]string, error) {
+func getEnvVars(arg starlark.Value, sep string) ([]string, error) {
 	if arg.Type() != "dict" {
 		return nil, fmt.Errorf("sh internal error: expected dict, got: %s", arg.Type())
 	}
-	d := arg.(*Dict)
+	d := arg.(*starlark.Dict)
 	res := []string{}
 	for _, k := range d.Keys() {
-		name, _ := AsString(k)
+		name, _ := starlark.AsString(k)
 		v, _, _ := d.Get(k)
 		vals := getDeepEnvVars(v, name, sep)
 		res = append(res, vals...)
@@ -55,25 +56,25 @@ func getEnvVars(arg Value, sep string) ([]string, error) {
 	return res, nil
 }
 
-func getDeepEnvVars(v Value, prefix, sep string) []string {
+func getDeepEnvVars(v starlark.Value, prefix, sep string) []string {
 	res := []string{}
 	switch v.Type() {
 	case "string":
-		val, _ := AsString(v)
+		val, _ := starlark.AsString(v)
 		res = []string{fmt.Sprintf("%s=%s", prefix, val)}
 	case "int", "float", "bool", "NoneType":
 		res = []string{fmt.Sprintf("%s=%s", prefix, v.String())}
 	case "dict":
-		d := v.(*Dict)
+		d := v.(*starlark.Dict)
 		for _, k := range d.Keys() {
-			name, _ := AsString(k)
+			name, _ := starlark.AsString(k)
 			newPrefix := fmt.Sprintf("%s%s%s", prefix, sep, name)
 			v, _, _ := d.Get(k)
 			res = append(res, getDeepEnvVars(v, newPrefix, sep)...)
 		}
 	case "tuple", "list", "set":
-		iter := Iterate(v)
-		var z Value
+		iter := starlark.Iterate(v)
+		var z starlark.Value
 		i := 0
 		for iter.Next(&z) {
 			i++
